# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wordpress/status:
    get:
      operationId: getWordpressStatus
      tags: [wordpress]
      summary: Get WordPress connection status for a silo
      parameters:
        - name: silo
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Silo"
      responses:
        "200":
          description: WordPress connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordpressStatus"
  /wordpress/publish:
    post:
      operationId: publishToWordpress
      tags: [wordpress]
      summary: Publish or draft an article to WordPress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordpressPublishInput"
      responses:
        "200":
          description: Publish result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WordpressPublishResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Approval required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /public-app/publish:
    post:
      operationId: publishToPublicApp
      tags: [public-app]
      summary: Publish an article to the public app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicAppPublishInput"
      responses:
        "200":
          description: Publish result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAppPublishResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Enums ───────────────────────────────────────────────────────────────
    Silo:
      type: string
      enum:
        - hiphophaven
        - raphaven
        - musichaven
        - sportshaven
        - fithaven
        - cannahaven
        - hmg
...
    PublishRequestStatus:
      type: string
      enum:
        - draft
        - publish

    # ─── Shared sub-schemas ───────────────────────────────────────────────────
...
    WordpressStatus:
      type: object
      properties:
        configured:
          type: boolean
        siteUrl:
          type: string
      required:
        - configured
    WordpressPublishResponse:
      type: object
      properties:
        link:
          type: string
        url:
          type: string
        status:
          type: string
        title:
          type: string
        postId:
          type: integer
...
    PublicAppRemote:
      type: object
      properties:
        url:
          type: string
        link:
          type: string
    PublicAppPublishResponse:
      type: object
      properties:
        url:
          type: string
        link:
          type: string
        status:
          type: string
        configured:
          type: boolean
        remote:
          $ref: "#/components/schemas/PublicAppRemote"

    # ─── Request body schemas (entity-named, never <OperationIdPascal>Body) ──
...
    WordpressPublishInput:
      type: object
      properties:
        silo:
          $ref: "#/components/schemas/Silo"
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        metaDescription:
          type: string
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/PublishRequestStatus"
        slug:
          type: string
        canonicalUrl:
          type: string
        videoEmbed:
          type: string
        featuredMediaId:
          type: integer
        overrideUrl:
          type: string
        overrideUser:
          type: string
        overridePassword:
          type: string
      required:
        - silo
        - title
        - content
    PublicAppPublishInput:
      type: object
      properties:
        silo:
          $ref: "#/components/schemas/Silo"
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        slug:
          type: string
        canonicalUrl:
          type: string
        videoEmbed:
          type: string
        featuredImageUrl:
          type: string
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required:
        - silo
        - title
        - content