# lib/api-zod/src/generated/types/publicAppPublishInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Silo } from './silo';

export interface PublicAppPublishInput {
  silo: Silo;
  title: string;
  content: string;
  excerpt?: string;
  slug?: string;
  canonicalUrl?: string;
  videoEmbed?: string;
  featuredImageUrl?: string;
  categories?: string[];
  tags?: string[];
}