# lib/api-zod/src/generated/types/wordpressPublishInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublishRequestStatus } from './publishRequestStatus';
import type { Silo } from './silo';

export interface WordpressPublishInput {
  silo: Silo;
  title: string;
  content: string;
  excerpt?: string;
  metaDescription?: string;
  categories?: string[];
  tags?: string[];
  status?: PublishRequestStatus;
  slug?: string;
  canonicalUrl?: string;
  videoEmbed?: string;
  featuredMediaId?: number;
  overrideUrl?: string;
  overrideUser?: string;
  overridePassword?: string;
}