# lib/api-zod/src/generated/types/wordpressPublishResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordpressPublishResponse {
  link?: string;
  url?: string;
  status?: string;
  title?: string;
  postId?: number;
}